# src/test/resources/contract/early-learning-api.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "早期学习困难儿童筛查与干预系统 HTTP API",
    "description": "Android 平板教师端与管理端的功能接口，共 39 个操作。依据《公共工程基础》九份资料、《技术方案》及用户确认口径，为待实现和联调的设计契约。\n\n成功响应 {code,message,data}；错误响应 {code,message,data:null,details?}。使用 Bearer Token 鉴权。标明 Idempotency-Key 的写入，同一操作重试沿用原值，防止重复创建或消耗凭证；同键不同输入返回 409 IDEMPOTENCY_CONFLICT。激活码/凭证只允许短时内存重放，失效返回 SENSITIVE_RESULT_EXPIRED，不重复生成。\n\nAI request_id 防止重复任务，input_revision 用于拒收旧输入的结果。评分标准按基本固定的一套服务端配置使用，不做自动更新、整场预检或服务端会话锁版。首次评分直接提交，服务端返回实际 rubric_version；后续同一评估/课堂沿用该版本即可。所有故事共用同一套评分规则，上新故事不触发评分标准更新。评分结果保存在本地 JSON，不自动重算或改写。转写与评分分开，文本须经教师确认。\n\n儿童业务数据留在本地；AI 音频、文本、图片及结果仅内存临时处理，不云端持久保存。官方内容单向下载，被引用文件不可删除。故事采用 AIScore v2，问答采用 QuestionAIScore；课程/材料必含故事叙述，字典无版本。\n\n调用顺序见 docs/api/业务流程与接口调用.md，导入说明见 docs/api/README.md。示例均为结构演示，不能作为真实内容或凭证使用。\n\n课程和评估材料发布改为 multipart 的 file 字段上传 ZIP，config.json 与素材放在包内。后端转换文件名后保存原冻结结构；现有详情、下载与单文件接口保留，接口总数仍为 39。\n\n刷新凭证不设到期时间，仅在刷新成功时轮换；上一枚凭证通过现有内存缓存保留 30 秒重试宽限期，窗口内返回同一组新凭证，不重复轮换或延长窗口。数据库仅保存当前 refresh_token_hash，不加字段或表。访问凭证仍有有效期。云端等待 ZIP 异步校验与文件准备全部成功后，在短事务中发布新版并禁用同编号旧版，再返回 201。平板独立等待全部资源下载和校验成功后，在本地短事务中启用新版并禁用旧版；任一步失败保留原可用状态。",
    "version": "1.6.0"
  },
...
  "paths": {
...
    "/admin/licenses": {
      "post": {
        "summary": "单个或批量生成激活码",
        "deprecated": false,
        "description": "count=1 为单个，一批最多 100 个，整批创建。原码只在成功响应和短时内存重放返回，服务端只持久保存哈希。同键不重复生成；原码无法重放时返回 409 SENSITIVE_RESULT_EXPIRED，并从原请求结果关联返回 details.license_ids，供逐个撤销。不得按最新列表猜本批 ID。",
        "operationId": "createLicenses",
        "tags": [
          "激活码"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "防止重复提交。客户端生成 UUID，同一次操作重试沿用；同键不同输入返回 409 IDEMPOTENCY_CONFLICT。",
            "required": true,
            "example": {},
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "count": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "生成数量。",
                    "maximum": 100
                  }
                },
                "required": [
                  "count"
                ]
              },
              "examples": {
                "structure": {
                  "value": {
                    "count": 1
                  },
                  "summary": "仅结构示例；不代表可用于生产的内容或凭证"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "创建成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/createLicensesResponse"
                },
                "example": {
                  "code": "OK",
                  "message": "成功",
                  "data": {
                    "items": [
                      {
                        "id": 1,
                        "activation_code": "EXAMPLE_ACTIVATION_CODE_NOT_VALID",
                        "status": "UNUSED"
                      }
                    ]
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "请求参数或编码不合法。"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "缺少凭证、凭证失效或账号密码不正确。"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "账号/激活码不可用、凭证类型错误或无资源权限。"
          },
          "409": {
            "description": "当前状态冲突或同一逻辑请求的输入已改变。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictResponse"
                },
                "example": {
                  "code": "SENSITIVE_RESULT_EXPIRED",
                  "message": "原码已无法重放，可按本次生成的记录 ID 撤销。",
                  "data": null,
                  "details": {
                    "license_ids": [
                      1
                    ]
                  }
                }
              }
            },
            "headers": {}
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests",
            "description": "请求频率或并发超过上限，请稍后重试。"
          },
          "500": {
            "$ref": "#/components/responses/InternalError",
            "description": "内部错误，请稍后重试。"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable",
            "description": "服务或必需依赖暂不可用。"
          }
        },
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "x-design-origin": "原文接口清单 + 用户确认修订 + 授权补充协议"
      },
      "get": {
        "summary": "查询激活码状态",
        "deprecated": false,
        "description": "按 id 降序分页。筛选条件同时满足；不提供原码、激活码哈希的查询或返回。",
        "operationId": "listLicenses",
        "tags": [
          "激活码"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "页码，从 1 开始。",
            "required": false,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "页码。",
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "每页条数，默认 20，最大 100。",
            "required": false,
            "example": "",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "每页条数。",
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "状态筛选。",
            "required": false,
            "example": {},
            "schema": {
              "$ref": "#/components/schemas/LicenseStatus"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "精确匹配绑定教师云端 ID。",
            "required": false,
            "example": {},
            "schema": {
              "$ref": "#/components/schemas/CloudId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/listLicensesResponse"
                },
                "example": {
                  "code": "OK",
                  "message": "成功",
                  "data": {
                    "items": [
                      {
                        "id": 1,
                        "user_id": null,
                        "status": "UNUSED",
                        "activated_at": null
                      }
                    ],
                    "page": 1,
                    "page_size": 20,
                    "total": 1
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "请求参数或编码不合法。"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "缺少凭证、凭证失效或账号密码不正确。"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "账号/激活码不可用、凭证类型错误或无资源权限。"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests",
            "description": "请求频率或并发超过上限，请稍后重试。"
          },
          "500": {
            "$ref": "#/components/responses/InternalError",
            "description": "内部错误，请稍后重试。"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable",
            "description": "服务或必需依赖暂不可用。"
          }
        },
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "x-design-origin": "原文接口清单 + 用户确认修订 + 授权补充协议"
      }
    },
    "/admin/licenses/{id}/revoke": {
      "post": {
        "summary": "撤销激活码",
        "deprecated": false,
        "description": "无请求体，路径 id 即激活码 ID。UNUSED/ACTIVE 均可撤销；ACTIVE 撤销与绑定教师禁用同一事务提交，同时阻断其云端访问和刷新。已 REVOKED 幂等返回。不可恢复或更换绑定，不影响本地离线业务。",
        "operationId": "revokeLicense",
        "tags": [
          "激活码"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "云端记录 ID。",
            "required": true,
            "example": {},
            "schema": {
              "$ref": "#/components/schemas/CloudId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作成功。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/revokeLicenseResponse"
                },
                "example": {
                  "code": "OK",
                  "message": "成功",
                  "data": {
                    "id": 1,
                    "user_id": 1,
                    "status": "REVOKED",
                    "activated_at": "2026-09-24T00:00:00.000Z"
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "$ref": "#/components/responses/BadRequest",
            "description": "请求参数或编码不合法。"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized",
            "description": "缺少凭证、凭证失效或账号密码不正确。"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden",
            "description": "账号/激活码不可用、凭证类型错误或无资源权限。"
          },
          "404": {
            "$ref": "#/components/responses/NotFound",
            "description": "资源不存在；其他账号的 AI 任务同样返回 TASK_NOT_FOUND。"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests",
            "description": "请求频率或并发超过上限，请稍后重试。"
          },
          "500": {
            "$ref": "#/components/responses/InternalError",
            "description": "内部错误，请稍后重试。"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable",
            "description": "服务或必需依赖暂不可用。"
          }
        },
        "security": [
          {
            "AdminBearer": []
          }
        ],
        "x-design-origin": "原文接口清单 + 用户确认修订 + 授权补充协议"
      }
    },
...
  },
  "components": {
    "schemas": {
      "CloudId": {
        "type": "integer",
        "minimum": 1,
        "description": "云端数据库 ID；不能作为本地 SQLite 外键直接保存。",
        "maximum": 9007199254740991
      },
      "Code": {
        "type": "string",
        "description": "稳定业务编号；不含路径分隔符。",
        "minLength": 1,
        "pattern": "^[A-Za-z0-9_-]+$",
        "maxLength": 128
      },
...
      "CloudFileCode": {
        "type": "string",
        "description": "官方资源编号；终身对应同一份字节，不复用。",
        "minLength": 1,
        "pattern": "^CF_[A-Za-z0-9_-]+$",
        "maxLength": 128
      },
...
      "Uuid": {
        "type": "string",
        "description": "客户端生成的随机 UUID；不得包含儿童身份信息。",
        "format": "uuid"
      },
...
      "ConflictResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "USERNAME_EXISTS",
              "LICENSE_UNAVAILABLE",
              "LICENSE_REVOKED",
              "CONTENT_VERSION_EXISTS",
              "RESOURCE_IN_USE",
              "IDEMPOTENCY_CONFLICT",
              "SENSITIVE_RESULT_EXPIRED",
              "TASK_RETRY_CONFLICT",
              "VERSION_CONFLICT",
              "RESOURCE_NOT_READY"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "ErrorDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "field_path": {
            "type": "string",
            "description": "第一个失败字段的 JSON Pointer；普通请求相对请求体根，ZIP 发布相对 config.json 根。路径/query 参数使用 /parameters/参数名，不嵌入用户文本。",
            "pattern": "^/"
          },
          "file_code": {
            "$ref": "#/components/schemas/CloudFileCode"
          },
          "rubric_item_code": {
            "$ref": "#/components/schemas/Code"
          },
          "current_version": {
            "type": "integer",
            "minimum": 1,
            "description": "语法修改发生版本冲突时的服务端当前版本。"
          },
          "limit": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "name": {
                "type": "string",
                "enum": [
                  "size_bytes",
                  "duration_ms",
                  "image_count",
                  "text_length"
                ],
                "description": "被触发限制的名称；text_length 按 UTF-16 code unit。"
              },
              "maximum": {
                "type": "integer",
                "minimum": 0,
                "description": "部署实际允许上限，单位由 name 指定。"
              }
            },
            "required": [
              "name",
              "maximum"
            ]
          },
          "license_ids": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CloudId"
            },
            "description": "仅生成激活码的 SENSITIVE_RESULT_EXPIRED 返回：该幂等请求实际创建的全部云端 id，可逐个调用既有 revoke；不含原码/哈希。",
            "minItems": 1,
            "maxItems": 100,
            "uniqueItems": true
          },
          "file_name": {
            "$ref": "#/components/schemas/PackageFileName"
          }
        },
        "description": "按错误类型提供必要的非敏感定位；不要求每项齐全，不回显儿童原话、凭证、输入值或整份请求。一次定位首个失败即可；无需遍历其他未获授权资源。",
        "minProperties": 1
      },
      "PackageFileName": {
        "type": "string",
        "minLength": 1,
        "maxLength": 128,
        "pattern": "^(?!\\.{1,2}$)[^<>:\"/\\\\|?*\\u0000-\\u001F\\u007F]+$",
        "description": "ZIP 根目录中的文件名，不带路径；首尾不得为空白。包内文件名忽略大小写不得重复，JSON 引用必须与实际文件名大小写完全一致。"
      },
...
      "BadRequestResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INVALID_REQUEST"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "UnauthorizedResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "TOKEN_MISSING",
              "TOKEN_INVALID",
              "TOKEN_EXPIRED",
              "INVALID_CREDENTIALS",
              "REFRESH_TOKEN_INVALID"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "ForbiddenResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ACCOUNT_DISABLED",
              "LICENSE_REVOKED",
              "AUTH_ROLE_MISMATCH",
              "RESOURCE_FORBIDDEN"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "NotFoundResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "RESOURCE_NOT_FOUND",
              "TASK_NOT_FOUND"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
      "TooManyRequestsResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "RATE_LIMITED"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "InternalErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "INTERNAL_ERROR"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "UnavailableResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "SERVICE_UNAVAILABLE",
              "RUBRIC_UNAVAILABLE",
              "DEPENDENCY_UNAVAILABLE"
            ],
            "description": "该 HTTP 状态下的稳定错误码。"
          },
          "message": {
            "type": "string",
            "description": "可展示的脱敏说明，不包含输入文本、密码或 Token。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "data": {
            "type": "object",
            "enum": [
              null
            ],
            "description": "必须为 null",
            "properties": {},
            "nullable": true
          },
          "details": {
            "$ref": "#/components/schemas/ErrorDetails"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "NullableTimestamp": {
        "type": "string",
        "description": "UTC ISO 8601 时间戳。",
        "format": "date-time",
        "nullable": true
      },
...
      "LicenseStatus": {
        "type": "string",
        "enum": [
          "UNUSED",
          "ACTIVE",
          "REVOKED"
        ],
        "description": "激活码状态；REVOKED 不可恢复。"
      },
...
      "License": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CloudId"
          },
          "user_id": {
            "type": "integer",
            "minimum": 1,
            "description": "云端数据库 ID；不能作为本地 SQLite 外键直接保存。",
            "maximum": 9007199254740991,
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/LicenseStatus"
          },
          "activated_at": {
            "$ref": "#/components/schemas/NullableTimestamp"
          }
        },
        "required": [
          "id",
          "user_id",
          "status",
          "activated_at"
        ]
      },
      "IssuedLicense": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "$ref": "#/components/schemas/CloudId"
          },
          "activation_code": {
            "type": "string",
            "description": "仅本次生成结果返回的可分发激活码。不得写日志。",
            "minLength": 1,
            "pattern": "\\S"
          },
          "status": {
            "type": "string",
            "enum": [
              "UNUSED"
            ],
            "description": "新生成状态。"
          }
        },
        "required": [
          "id",
          "activation_code",
          "status"
        ]
      },
...
      "createLicensesResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "OK"
            ],
            "description": "业务请求受理或读取成功；异步结果须进一步检查任务 stage。"
          },
          "message": {
            "type": "string",
            "description": "脱敏说明。"
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/IssuedLicense"
                },
                "description": "仅本次生成的可分发激活码。",
                "minItems": 1,
                "maxItems": 100
              }
            },
            "required": [
              "items"
            ]
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "LicensePage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/License"
            },
            "description": "当前页记录。"
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "description": "当前页码。"
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "description": "每页条数。",
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "过滤后的总条数。"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total"
        ]
      },
      "listLicensesResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "OK"
            ],
            "description": "业务请求受理或读取成功；异步结果须进一步检查任务 stage。"
          },
          "message": {
            "type": "string",
            "description": "脱敏说明。"
          },
          "data": {
            "$ref": "#/components/schemas/LicensePage"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
      "revokeLicenseResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "OK"
            ],
            "description": "业务请求受理或读取成功；异步结果须进一步检查任务 stage。"
          },
          "message": {
            "type": "string",
            "description": "脱敏说明。"
          },
          "data": {
            "$ref": "#/components/schemas/License"
          }
        },
        "required": [
          "code",
          "message",
          "data"
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数或编码不合法。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BadRequestResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "INVALID_REQUEST",
                  "message": "请求参数或编码不合法。",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "缺少凭证、凭证失效或账号密码不正确。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UnauthorizedResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "TOKEN_MISSING",
                  "message": "缺少凭证、凭证失效或账号密码不正确。",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "账号/激活码不可用、凭证类型错误或无资源权限。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ForbiddenResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "ACCOUNT_DISABLED",
                  "message": "账号/激活码不可用、凭证类型错误或无资源权限。",
                  "data": null
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在；其他账号的 AI 任务同样返回 TASK_NOT_FOUND。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NotFoundResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "RESOURCE_NOT_FOUND",
                  "message": "资源不存在；其他账号的 AI 任务同样返回 TASK_NOT_FOUND。",
                  "data": null
                }
              }
            }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "请求频率或并发超过上限，请稍后重试。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/TooManyRequestsResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "RATE_LIMITED",
                  "message": "请求频率或并发超过上限，请稍后重试。",
                  "data": null
                }
              }
            }
          }
        }
      },
      "InternalError": {
        "description": "内部错误，请稍后重试。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/InternalErrorResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "INTERNAL_ERROR",
                  "message": "内部错误，请稍后重试。",
                  "data": null
                }
              }
            }
          }
        }
      },
      "Unavailable": {
        "description": "服务或必需依赖暂不可用。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/UnavailableResponse"
            },
            "examples": {
              "成功示例": {
                "summary": "成功示例",
                "value": {
                  "code": "SERVICE_UNAVAILABLE",
                  "message": "服务或必需依赖暂不可用。",
                  "data": null
                }
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "管理员登录返回的 token。每次请求校验账号仍为 ACTIVE；不能以教师 Token 替代。"
      },
      "TeacherBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "注册或刷新取得的 access_token；每次请求校验教师状态为 1、绑定激活码为 ACTIVE。禁用不影响本地登录。"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "TeacherBearer": []
    }
  ]
}